Various changes
Essence of the firelord was flagged as the same classes Eye of Sulfuras (Classes that could wield 2H maces)
This has now been changed to reflect the classes marked on Bindings.

Recipe: Shadow Oil wasn't sold by Vendor-Tron 1000, it is no longer listed under it's stock and is now listed under Thousand Needles and Duskwood as intended.

Your place in the world is now flagged as a breadcrumb.

Salve by disenchanting quest IDs for horde were mixed up, this has been changed.

Journey to Crossroads has now been flagged as a breadcrumb.